Move stories to component folders
