+++ src/components/account/stories.js
@@ -1,18 +1,8 @@
 import React from 'react';
 
 import { storiesOf } from '@storybook/react';
-import { action } from '@storybook/addon-actions';
-// import { linkTo } from '@storybook/addon-links';
-
-import FormattedNumber from '../src/components/formattedNumber';
-import Account from '../src/components/account/accountComponent';
-import Address from '../src/components/account/address';
-import Dialog from '../src/components/dialog/dialogElement';
-
-storiesOf('FormattedNumber', module)
-  .add('with val', () => (
-    <FormattedNumber val="-3.1415926535" />
-  ));
+import Account from './accountComponent';
+import Address from './address';
 
 storiesOf('Account', module)
   .add('delegate', () => (
@@ -50,18 +40,5 @@ storiesOf('Address', module)
     <Address
       isDelegate={false}
       address="9396639332432599292L"
-    />
-  ));
-
-const dialogContent = () => (<div>Hello</div>);
-
-storiesOf('Dialog', module)
-  .add('default', () => (
-    <Dialog
-      dialog={{
-        title: 'Title',
-        childComponent: dialogContent,
-      }}
-      onCancelClick={ action('onCancelClick') }
     />
   ));
