+++ src/components/formattedNumber/stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { storiesOf } from '@storybook/react';
+
+import FormattedNumber from './';
+
+storiesOf('FormattedNumber', module)
+  .add('with val', () => (
+    <FormattedNumber val="-3.1415926535" />
+  ));
